1. added scipy to the setup config file 2. places RelPerm class in it's own py file to reduce lines of code in dataframe.py 3. added necessary "assert" keyword to line 127 of test_relperm.py
