+++ setup.cfg
@@ -30,6 +30,7 @@ install_requires =
     h5py
     lxml
     lasio
+    scipy
 
 [options.package_data]
 # Include resqpy/olio/data/*.json
